refactor: split controller.go into focused files by concern
Split the 995-line controller.go into four files organized by responsibility:

- controller.go (~195 lines): types, interfaces, Controller struct, New(), Run()
- handlers.go (~205 lines): event handler registration, worker loop
- reconcile.go (~450 lines): event processing, deployment recording, workload
  resolution, cache existence checks
- pod.go (~195 lines): pod utility functions, informer factory

Extract registerEventHandlers() from New() as a Controller method and
newAPIClient() as a package-level helper to simplify the constructor.

Fix a pre-existing bug in DeleteFunc where MetaNamespaceKeyFunc(obj) was
called on the original obj which may be a DeletedFinalStateUnknown tombstone
(causing silent event drops). Now uses the extracted pod instead.

No other behavioral changes — pure structural refactoring.
